refactor & fix plugins not getting argument object
