+++ src/args.ts
@@ -1,5 +1,5 @@
 import yargs from "yargs";
-import * as logger from "./logger/index";
+import * as logger from "./logger";
 
 const argv = yargs.boolean("commit-import").argv;
 logger.log(argv);
