+++ src/fs/index.ts
@@ -1,0 +1,3 @@
+import { lstatSync } from "fs";
+
+export const isDirectory = (path: string) => lstatSync(path).isDirectory();
