+++ src/graphql/mutations/label.ts
@@ -5,7 +5,7 @@ import Scene from "../../types/scene";
 import Image from "../../types/image";
 import { Dictionary } from "../../types/utility";
 import { stripStr } from "../../extractor";
-import * as logger from "../../logger/index";
+import * as logger from "../../logger";
 
 type ILabelUpdateOpts = Partial<{
   name: string;
