+++ src/graphql/resolvers/search/movie.ts
@@ -1,6 +1,6 @@
 import Movie from "../../../types/movie";
 import extractQueryOptions, { SortTarget } from "../../../query_extractor";
-import * as logger from "../../../logger/index";
+import * as logger from "../../../logger";
 import * as search from "../../../search/index";
 import { IMovieSearchDoc } from "../../../search/movie";
