+++ src/graphql/resolvers/search/scene.ts
@@ -1,6 +1,6 @@
 import Scene from "../../../types/scene";
 import extractQueryOptions, { SortTarget } from "../../../query_extractor";
-import * as logger from "../../../logger/index";
+import * as logger from "../../../logger";
 import * as search from "../../../search/index";
 import { ISceneSearchDoc } from "../../../search/scene";
