+++ src/graphql/resolvers/search/studio.ts
@@ -1,6 +1,6 @@
 import Studio from "../../../types/studio";
 import extractQueryOptions, { SortTarget } from "../../../query_extractor";
-import * as logger from "../../../logger/index";
+import * as logger from "../../../logger";
 import * as search from "../../../search/index";
 import { IStudioSearchDoc } from "../../../search/studio";
