+++ src/plugin_events/actor.ts
@@ -1,5 +1,5 @@
 import Actor from "../types/actor";
-import { runPluginsSerial } from "../plugins";
+import { runPluginsSerial } from "../plugins/index";
 import { libraryPath } from "../types/utility";
 import { extractLabels, extractFields } from "../extractor";
 import { getConfig } from "../config";
@@ -7,7 +7,7 @@ import { extname } from "path";
 import { downloadFile } from "../ffmpeg-download";
 import Image from "../types/image";
 import * as database from "../database/index";
-import * as logger from "../logger/index";
+import * as logger from "../logger";
 import { indices } from "../search/index";
 import { createImageSearchDoc } from "../search/image";
 import Label from "../types/label";
