+++ src/plugin_events/movie.ts
@@ -1,4 +1,4 @@
-import { runPluginsSerial } from "../plugins";
+import { runPluginsSerial } from "../plugins/index";
 import { libraryPath } from "../types/utility";
 import { extractLabels, extractFields } from "../extractor";
 import { getConfig } from "../config";
@@ -6,7 +6,7 @@ import { extname } from "path";
 import { downloadFile } from "../ffmpeg-download";
 import Image from "../types/image";
 import * as database from "../database/index";
-import * as logger from "../logger/index";
+import * as logger from "../logger";
 import { indices } from "../search/index";
 import { createImageSearchDoc } from "../search/image";
 import Label from "../types/label";
