+++ src/search/image.ts
@@ -1,7 +1,7 @@
 import { SearchIndex } from "./engine";
 import Image from "../types/image";
 import { tokenizeNames, tokenize } from "./tokenize";
-import * as log from "../logger/index";
+import * as log from "../logger";
 import { memorySizeOf } from "../mem";
 import ora from "ora";
 import { inspect } from "util";
