+++ src/types/utility.ts
@@ -1,6 +1,6 @@
 import { getConfig } from "../config/index";
 import * as path from "path";
-import * as logger from "../logger/index";
+import * as logger from "../logger";
 
 export interface Dictionary<T> {
   [key: string]: T;
